services/collector/requirements.txt
ping3==4.0.7
psutil==6.0.0
PyYAML==6.0.2
requests==2.32.3
SQLAlchemy==2.0.32
pytest==8.1.1